sweetalert2 added

+++ app/scripts/controllers/post.js
@@ -8,8 +8,8 @@
  * Controller of the angularBoilerplate
  */
 angular.module('angularBoilerplate')
-  .controller('PostCtrl', ['$scope', 'baseService',
-  function ($scope, baseService) {
+  .controller('PostCtrl', ['$scope', 'swangular', 'baseService',
+  function ($scope, swangular, baseService) {
     $scope.url = 'http://127.0.0.1:3000/';
 
     $scope.getAll = function(){
@@ -17,17 +17,19 @@ angular.module('angularBoilerplate')
     };
 
     function updateData(){
+      $scope.post = '';
       $scope.refreshing = true;
       baseService.list($scope.url)
       .success(function (response) {
         if(response.status === "SUCCESS") {
           $scope.posts = response.data;
+          $scope.refreshing = !$scope.refreshing;
         }
       })
       .error(function (response, status, headers, config, scope) {
         console.log(response);
+        $scope.refreshing = !$scope.refreshing;
       })
-      $scope.refreshing = !$scope.refreshing;
     }
 
     $scope.new = function(){
@@ -62,14 +64,14 @@ angular.module('angularBoilerplate')
       baseService.create($scope.url + 'create', data)
       .success(function (response) {
         if (response.status === "SUCCESS") {
-          $scope.message = response.message;
           updateData();
+          swangular.success('Post created.')
         }
       })
       .error(function (response, status, headers, config, scope) {
         console.log(response);
+        $scope.refreshing = !$scope.refreshing;
       })
-      $scope.refreshing = !$scope.refreshing;
     };
 
     $scope.updatePost = function(){
@@ -84,28 +86,47 @@ angular.module('angularBoilerplate')
         if (response.status === "SUCCESS") {
           $scope.posts = response.data;
           updateData();
+          swangular.success('Post updated.')
         }
       })
       .error(function (response, status, headers, config, scope) {
         console.log(response);
+        $scope.refreshing = !$scope.refreshing;
       })
-      $scope.refreshing = !$scope.refreshing;
     };
 
     $scope.destroyPost = function(id){
       $scope.refreshing = true;
-      baseService.destroy($scope.url + 'delete/', id)
-      .success(function (response) {
-        if (response.status === "SUCCESS") {
-          updateData();
-          $scope.post = '';
+      swangular.swal({
+        title: 'Are you sure?',
+        text: 'Delete post? ',
+        type: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, i\'m sure!'
+      }).then((result) => {
+        if (result.value) {
+          baseService.destroy($scope.url + 'delete/', id)
+          .success(function (response) {
+            if (response.status === "SUCCESS") {
+              updateData();
+              swangular.success('Post removed.');
+            }
+          })
+          .error(function (response, status, headers, config, scope) {
+            console.log(response);
+            swangular.warning(response);
+            $scope.refreshing = !$scope.refreshing;
+          })
         }
       })
-      .error(function (response, status, headers, config, scope) {
-        console.log(response);
-      })
+
+
+
       $scope.refreshing = !$scope.refreshing;
-    };
+    }
+
 
     // executado ao fechar o modal
     $scope.cancel = function(){
